Split up functions into modules, tests.
